requirements.txt
gitpython
transformers 
datasets
sentence-transformers
transformers[torch]


# util - metric
seqeval
evaluate

pytorch-crf